move neo-doc-loader to separate dir
Also removed old transmute doc loader
(was called document-loader previously)
