+++ credentials/2021-07-exchange-of-enegy-blocks/human-identity/human-identity.test.ts
@@ -2,6 +2,7 @@ import schema from './human-identity-schema.json'
 import credential from './human-identity-credential.json'
 import { verifyCredentialSubjectSchema } from '../../../test/verify-credential-subject-schema'
 import { issueAndVerify } from '../../../test/issue-and-verify'
+import { documentLoader } from '../../../test/document-loader'
 
 describe("AssetSpec", () => {
   test("Credential should match JSON Schema", async () => {
@@ -9,6 +10,6 @@ describe("AssetSpec", () => {
   })
 
   test("Credential can be issued and verified", async () => {
-    await issueAndVerify(credential)
+    await issueAndVerify(credential, documentLoader)
   })
 })
