+++ credentials/2021-07-exchange-of-enegy-blocks/link-company-to-human/link-company-to-human.test.ts
@@ -1,6 +1,7 @@
 // import schema from './link-company-to-human.json'
 import credential from './link-company-to-human.json'
 import { issueAndVerify } from '../../../test/issue-and-verify'
+import { documentLoader } from '../../../test/document-loader'
 
 describe("Link Company To Human", () => {
   // test("Credential should match JSON Schema", async () => {
@@ -8,6 +9,6 @@ describe("Link Company To Human", () => {
   // })
 
   test("Credential can be issued and verified", async () => {
-    await issueAndVerify(credential)
+    await issueAndVerify(credential, documentLoader)
   })
 })
