+++ credentials/2021-07-exchange-of-enegy-blocks/terms-and-conditions/terms-and-conditions.test.ts
@@ -1,5 +1,6 @@
 import credential from './terms-and-conditions-credential.json'
 import { issueAndVerify } from '../../../test/issue-and-verify'
+import { documentLoader } from '../../../test/document-loader'
 
 describe("Terms And Conditions", () => {
   // test("Credential should match JSON Schema", async () => {
@@ -7,6 +8,6 @@ describe("Terms And Conditions", () => {
   // })
 
   test("Credential can be issued and verified", async () => {
-    await issueAndVerify(credential)
+    await issueAndVerify(credential, documentLoader)
   })
 })
