+++ credentials/2021-07-exchange-of-enegy-blocks/time-series/time-series.test.ts
@@ -2,6 +2,7 @@ import schema from './time-series-schema.json'
 import credential from './time-series-credential.json'
 import { verifyCredentialSubjectSchema } from '../../../test/verify-credential-subject-schema'
 import { issueAndVerify } from '../../../test/issue-and-verify'
+import { documentLoader } from '../../../test/document-loader'
 
 describe("Time Series", () => {
   test("Credential should match JSON Schema", async () => {
@@ -9,6 +10,6 @@ describe("Time Series", () => {
   })
 
   test("Credential can be issued and verified", async () => {
-    await issueAndVerify(credential)
+    await issueAndVerify(credential, documentLoader)
   })
 })
