+++ credentials/2022-neo/device-info/device-info.test.ts
@@ -1,9 +1,7 @@
-import { check } from '@transmute/jsonld-schema'
-import { Options, expand } from 'jsonld';
+import { Options, expand } from 'jsonld'
 import credential from './device-info-credential.json'
-import { verifyCredentialSubjectSchema } from '../../../test/verify-credential-subject-schema'
 import { issueAndVerify } from '../../../test/issue-and-verify'
-import { documentLoader } from '../../../test/bloomDocumentLoader'
+import { digitalBazaarDocumentLoader, transmuteDocumentLoader } from '../neo-document-loader'
 
 describe("Device Info", () => {
   // test("Credential should match JSON Schema", async () => {
@@ -11,16 +9,14 @@ describe("Device Info", () => {
   // })
 
   test("Credential can be issued and verified", async () => {
-    await issueAndVerify(credential)
+    await issueAndVerify(credential, transmuteDocumentLoader)
   })
 
-  test("Credential is valid JSON-LD", async () => {
-    const result = await check({input: credential, documentLoader })
-    expect(result.ok).toBeTruthy()
-    // const options: Options.Expand = {
-    //   documentLoader: documentLoader
-    // }
-    // const expanded = await expand(credential, options)
-    // console.log(expanded)
+  test("Credential expands as expected", async () => {
+    const options: Options.Expand = {
+      documentLoader: digitalBazaarDocumentLoader
+    }
+    const expanded = await expand(credential, options)
+    console.log(expanded)
   })
 })
